+++ src/analysis/tax_transfer_funcs/benefits.py
@@ -1,126 +1,6 @@
 import numpy as np
 import pandas as pd
 
-from src.analysis.tax_transfer_funcs.taxes import soli_formula
-
-
-def ui(df_row, tb):
-    """Return the Unemployment Benefit based on
-    employment status and income from previous years.
-
-    """
-
-    ui = pd.DataFrame(index=df_row.index)
-
-    # ui["m_alg1_soep"] = df["alg_soep"].fillna(0)
-
-    ui["alg_entgelt"] = _alg_entgelt(df_row, tb)
-
-    ui["eligible"] = check_eligibility_alg(df_row)
-
-    ui["m_alg1"] = 0
-    ui.loc[ui["eligible"], "m_alg1"] = ui["alg_entgelt"] * np.select(
-        [df_row["child_num_tu"] == 0, df_row["child_num_tu"] > 0],
-        [tb["agsatz0"], tb["agsatz1"]],
-    )
-
-    return ui["m_alg1"]
-
-
-def _alg_entgelt(df_row, tb):
-    """ Calculating the claim for the Arbeitslosengeldgeld, depending on the current
-    wage."""
-    if df_row["east"].iloc[0]:
-        westost = "o"
-    else:
-        westost = "w"
-    # Relevant wage is capped at the contribution thresholds
-    alg_wage = np.minimum(tb["rvmaxek" + westost], df_row["m_wage_l1"])
-
-    # We need to deduct lump-sum amounts for contributions, taxes and soli
-    alg_ssc = tb["alg1_abz"] * alg_wage
-    # assume west germany for this particular calculation
-    # df['east'] = False
-    # Fictive taxes (Lohnsteuer) are approximated by applying the wage to the tax tariff
-    alg_tax = tb["tax_schedule"](12 * alg_wage - tb["werbung"], tb)
-
-    alg_soli = soli_formula(alg_tax, tb)
-
-    return np.maximum(0, alg_wage - alg_ssc - alg_tax / 12 - alg_soli / 12)
-
-
-def check_eligibility_alg(df_row):
-    """Checking eligibility, depending on the months worked beforehand, the age and
-    other variables.."""
-    # Months of unemployment beforehand.
-    mts_ue = df_row["months_ue"] + df_row["months_ue_l1"] + df_row["months_ue_l2"]
-    # BENEFIT AMOUNT
-    # Check Eligiblity.
-    # Then different rates for parent and non-parents
-    # Take into account actual wages
-    # there are different replacement rates depending on presence of children
-    return (
-        mts_ue.between(1, 12)
-        & (df_row["age"] < 65)
-        & (df_row["m_pensions"] == 0)
-        & (df_row["w_hours"] < 15)
-    )
-
-
-def uhv(df, tb):
-    """
-    Since 2017, the receipt of this
-    UHV has been extended substantially and needs to be taken into account, since it's
-    dominant to other transfers, i.e. single parents 'have to' apply for it.
-    """
-    if tb["yr"] >= 2017:
-        return uhv_since_2017(df, tb)
-    else:
-        return 0
-
-
-def uhv_since_2017(df, tb):
-    """ Advanced Alimony Payment / Unterhaltsvorschuss (UHV)
-
-        In Germany, Single Parents get alimony payments for themselves and for their
-        child from the ex partner. If the ex partner is not able to pay the child
-        alimony, the government pays the child alimony to the mother (or the father, if
-        he has the kids)
-
-        returns:
-            uhv (pd.Series): Alimony Payment on individual level
-        """
-    # Benefit amount depends on parameters M (rent) and Y (income) (§19 WoGG)
-    # Calculate them on the level of the tax unit
-    uhv = pd.DataFrame(index=df.index.copy())
-
-    uhv["uhv"] = 0
-    # Amounts depend on age
-    uhv.loc[df["age"].between(0, 5) & df["alleinerz"], "uhv"] = tb["uhv5"]
-    uhv.loc[df["age"].between(6, 11) & df["alleinerz"], "uhv"] = tb["uhv11"]
-    # Older kids get it only if the parent has income > 600€
-    uhv["uhv_inc_tu"] = (
-        df[
-            [
-                "m_wage",
-                "m_transfers",
-                "m_self",
-                "m_vermiet",
-                "m_kapinc",
-                "m_pensions",
-                "m_alg1",
-            ]
-        ]
-        .sum()
-        .sum()
-    )
-    uhv.loc[
-        (df["age"].between(12, 17)) & (df["alleinerz"]) & (uhv["uhv_inc_tu"] > 600),
-        "uhv",
-    ] = tb["uhv17"]
-    # TODO: Check against actual transfers
-    return uhv["uhv"]
-
 
 def benefit_priority(df, tb):
     """There are three main transfers for working-age people:
